Updated models and test fixtures
Updated the models to reflect the changes to the SHR grammar.

Updated the test fixtures to reflect the changes to the SHR grammar, and updated the tests to correctly reference the new fixtures.  Did NOT update test logic to use new models -- it still operates against old model definitions.  Did NOT update any translator logic either.