Make shr-models and bunyan peer dependencies
Due to the nature of the SHR tools, the entire chain must be guaranteed to be using the same instance of the shr-models library.  This has caused issues in the past and required every component of the chain to be updated and re-published whenever shr-models changed in any way.  A better approach to this is to declare shr-models as a peer dependency in these libraries and have SHR CLI provide the single instance of shr-models.

Bunyan is used in a similar way, and thus SHR CLI is also responsible for providing that library.

This commit also updates shr-models to 5.2.1.