Merge constraints down to tail of path
In order to collect all of the possible VS constraints anywhere in the path, merge all constraints down to the end.

This heavily steals code from shr-fhir-export that should eventually be centralized to shr-models.  In addition, the code could probably be pruned more (may do more than we need) -- but for now... it works!